pyproject.toml
[project]
name = "chatgpt-clone"
version = "0.1.0"
description = "Simple ChatGPT clone using LangGraph and Claude API"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "anthropic>=0.18.0",
    "langchain-anthropic>=0.1.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"

[tool.hatch.build.targets.wheel]
packages = ["src/chatgpt_clone"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"